Downloads 2 latest Windows files; Auto install
Downloads the latest 2 Windows files which will be the ActiveX version and the 3rd Party Browser version. Skips the download if the file exists already. Automatically calls the msiexec installer to install the file upon download.

+++ get_adobe_flash_updaters.py
@@ -2,15 +2,17 @@
 # Update to the latest Adobe Flash players
 # Limited to Windows right now though easy to change
 
-import sys, urllib2, subprocess
+import os
+import sys
+import urllib2
+import subprocess
 from BeautifulSoup import BeautifulSoup
 
-# -u commandline param = download update to this program
 if len(sys.argv) == 2 and sys.argv[1] == '-u':
-    url_opener = urllib2.urlopen('http://www.somedomain_somewhere.com/upd_flash.zip')
+    url_opener = urllib2.urlopen('http://www.some_domain.com/flashu/g_flash.zip')
     data = url_opener.read()
-    open('upd_flash.zip', "wb").write(data)
-    print 'Downloaded upd_flash.zip'
+    open('g_flash.zip', "wb").write(data)
+    print 'Downloaded g_flash.zip'
     sys.exit()
 
 url = 'http://www.adobe.com/products/flashplayer/distribution3.html'
@@ -21,17 +23,25 @@ html = conn.read()
 soup = BeautifulSoup(html)
 links = soup.findAll('a')
 
+download_counter = 0
 for tag in links:
     link = tag.get('href',None)
     if link != None and link.find('.msi') > -1 and link.find('/download') > -1:
-        print "Attempting download of " + link
         try:
-            url_opener = urllib2.urlopen(link)
-            data = url_opener.read()
             strfilename = link.rsplit('/', 1)[1]
-            open(strfilename, "wb").write(data)
-            print 'Downloaded ' + link.rsplit('/', 1)[1]
-            subprocess.call(strfilename)
+            if not os.path.exists(strfilename):
+                print "Attempting download of " + link
+                url_opener = urllib2.urlopen(link)
+                data = url_opener.read()
+                open(strfilename, "wb").write(data)
+                print 'Downloaded ' + link.rsplit('/', 1)[1]
+            else:
+                print strfilename, "already exists. skipping download."
+            subprocess.call("msiexec /quiet /passive /i " + strfilename)
+            download_counter += 1
+            if download_counter == 2:
+                print "Downloaded latest two .msi files..."
+                break
         except urllib2.HTTPError, val:
             if '404' in str(val): 
                 print "Couldn't download (404) " + url
